backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.9.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-jazzmin==3.0.1
django-swagger-ui==0.1.11
djangorestframework==3.15.2
djangorestframework-gis==1.2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
